pyproject.toml
[tool.poetry]
name = 'tevmc'
packages = [{include='tevmc', from='src'}]
version = '2.0.0-rc1'
description = 'Telos EVM Docker controller / framework'
authors = ['Guillermo Rodriguez <guillermo@telos.net>']
readme = 'README.md'

[tool.poetry.dependencies]
python = '>=3.10,<4'
click = '^8.1.7'
py-leap = {git = 'https://github.com/guilledk/py-leap', rev = 'v0.1a20'}
dockerstack = {git = 'https://github.com/guilledk/dockerstack', rev = 'v1.0.0-rc16'}
deepmerge = "^1.1.1"
beautifulsoup4 = "^4.12.3"
requests-unixsocket = "^0.3.0"
rlp = "^4.0.0"
web3 = "^6.15.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"
pdbp = "^1.5.0"

[build-system]
requires = ['poetry-core']
build-backend = 'poetry.core.masonry.api'

[tool.poetry.scripts]
tevmc = 'tevmc.cli:cli'

[build]
include = ['templates/**/*']

[tool.pytest.ini_options]
addopts = "-s"
markers = [
    "stack_config: set launch options for test stack"
]
